Move Lib-BDD parser out of the Apply benchmark for reuse
